requirements.txt
Flask=0.10.1
Pillow==2.2.2


